[공지사항] 헤더의 공지사항 클릭 시 상세페이지 구현
1. 공지사항 상세페이지 구현
   - 공지사항 모달 NoticeDetail로 변경

2. 헤더 공지사항 상세페이지에서 수정 버튼 클릭 시 모달 꺼짐 현상 발생
    - GridModal의 button type="button" 추가, setIsEdit(true) 주석처리로 해결
    - 원인 파악 못함.
